small changes like move or delete unnecessary folders
