+++ corpGoodsApp/pom.xml
@@ -4,7 +4,7 @@
   <groupId>bsu.fpmi.discounts</groupId>
   <artifactId>corpGoodsApp</artifactId>
   <packaging>war</packaging>
-  <version>1.0-SNAPSHOT</version>
+  <version>1.0-SNAPHOT</version>
   <name>corpGoodsApp Maven Webapp</name>
   <url>http://maven.apache.org</url>
   <dependencies>
@@ -18,6 +18,11 @@
       <groupId>javax.servlet</groupId>
       <artifactId>javax.servlet-api</artifactId>
       <version>4.0.1</version>
+    </dependency>
+    <dependency>
+      <groupId>com.fasterxml.jackson.core</groupId>
+      <artifactId>jackson-databind</artifactId>
+      <version>2.12.2</version>
     </dependency>
   </dependencies>
   <build>
