+++ corpGoodsApp/src/main/java/com/task5/servlets/filter.java
@@ -1,4 +1,4 @@
-package com.example.WebApplication;
+package com.task5.servlets;
 
 import javax.servlet.*;
 import javax.servlet.annotation.WebFilter;
