+++ corpGoodsApp/src/main/java/com/task5/servlets/nameServlet.java
@@ -1,4 +1,4 @@
-package com.example.WebApplication;
+package com.task5.servlets;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
