--- corpGoodsApp/src/main/webapp/index.jsp
@@ -1,5 +1,0 @@
-<html>
-<body>
-<h1>Java Web Application</h1>
-</body>
-</html>
